+++ src/app/components/Error404Page/Error404Page.jsx
@@ -1,10 +1,17 @@
 import React from 'react';
+import config from '../../../../appConfig';
 
 const Error404Page = () => (
   <div className="error-container">
     <h1 className="error-title">404 | Not Found</h1>
     <p>
       We’re sorry. The page you were looking for doesn’t exist.
+    </p>
+    <p>
+      Ready to discover your next great read? Browse and filter our&nbsp;
+      <a href={`${config.baseUrl}/staff-picks`}>
+        Staff Picks.
+      </a>
     </p>
     <p>
       Need help or have a question?&nbsp;